requirements.txt
Django==4.2.7 #for mysql and django rest framework
djangorestframework==3.14.0 
mysqlclient==2.2.0
opencv-python==4.8.1.78
python-doctr[torch]
pdf2image==1.16.3
Pillow==10.1.0
numpy==1.26.2
django-cors-headers==4.3.1
python-dotenv==1.0.0
pymupdf==1.23.5
